Move API tag sorting to REST settings.py
